Get travel advise from one station to another station

This is equivalent to the NS reisplanner, you give in a from and to station time, departure or arrival, and optionally if you have a NS year card (has effect on some travels). You can also specify how many trips before and after the chosen time you want to call.

travel_advise(fromStation, toStation, dateTime = NULL, departure = TRUE,
  yearCard = FALSE, hslAllowed = FALSE, previousAdvices = 4,
  nextAdvices = 4)

Arguments

fromStation
the station to start from, for instance "Rotterdam Centraal"

toStation
the station to end, for instance "Utrecht Centraal"

dateTime
defaults to current time, but you can use a different one: f.i. 2012-02-21T15:50, You can also use the datetime() function

departure
is the datetime the start or end time? do you want to depart on that date or arrive, defaults to departure

yearCard
if you have a NS year card (jaarabonnement) some trips will be different

hslAllowed
use of the highspeed train

previousAdvices
how many advices do you want before the time

nextAdvices
how many advises do you want after

Details

Some things to consider: station names need to be in Dutch and the NS webservice also accepts shortened versions: "Utrecht Centraal" and "ut" is apparently the same. Station names can be found with the stationfinder call.

Although the documentation https://www.ns.nl/en/travel-information/ns-api/documentation-travel-recommendations.html is in english, the returns are all in Dutch. And I keep the results in English.